consumer/requirements.txt
psycopg2-binary
redis
protobuf